No longer require the hazmat feature to use ed448
