+++ crypto/ed448/Cargo.toml
@@ -28,6 +28,3 @@ dalek-ff-group = { path = "../dalek-ff-group", version = "^0.1.2" }
 
 [dev-dependencies]
 hex = "0.4"
-
-[features]
-hazmat = []
