--- crypto/ed448/src/hazmat.rs
@@ -1,6 +1,0 @@
-pub(crate) use dalek_ff_group::math;
-
-mod backend;
-pub mod scalar;
-pub mod field;
-pub mod point;
