--- crypto/ed448/src/hazmat/backend.rs
@@ -1,9 +1,0 @@
-pub(crate) mod constant_time;
-
-pub mod scalar {
-  pub use crate::hazmat::backend::constant_time::scalar::*;
-}
-
-pub mod field {
-  pub use crate::hazmat::backend::constant_time::field::*;
-}
