+++ crypto/ed448/src/lib.rs
@@ -1,6 +1,6 @@
 #![no_std]
 
-#[cfg(feature = "hazmat")]
-mod hazmat;
-#[cfg(feature = "hazmat")]
-pub use hazmat::*;
+mod backend;
+pub mod scalar;
+pub mod field;
+pub mod point;
